Update to LLVM 3.6
Change the obfi code to use the new JIT API, so that it compiles with
LLVM 3.6.

As this now requires using ctypes, add it as a dependency in the _tags.
Change the _tags syntax to -use-ocamlfind while we're at it and remove
the lines in myocamlbuild.ml that are no longer necessary as a result.