Determine file names of results properly
Instead of deriving the name of the generated bitcode file by appending
"c" to the source file name (based on the assumption that it ends in
".b"), properly derive the name by replacing the source file's extension
(whichever it may be) with ".bc" or just appending ".bc" if it does not
have an extension.

Since this makes it possible to use source extensions other than
.b (without messing up the file names of the bitcode files), use ".bf"
as the extension in the example usage in the readme.

Also remove the paragraph explaining the old behavior from the readme.